requirements.txt
selenium==4.18.1
beautifulsoup4==4.12.3
requests==2.31.0 
google-generativeai>=0.3.0
playwright>=1.43.0
